refactor: make EqualizerManager() self-contained
